Lots of Changes
Added a new blockout for the new maze blockout.  Mesh is not smoothed
yet so some of the meshes are mis placed. Added maze blockout mesh,
jumping rocks, and mushrooms.  Added textures for amulet, jumping rocks,
and mushrooms.  Chenged the water texture so we can see them.  Added a
jump, changed the sprint, and changed the camera height to the player
character. Added rocks to the starting area.