Added VFX and Level Blueprint
Created a simple level Blueprint to prevent the player from going passed
the pound.

You can now go to the stone, press E and see the collision destroyed
(Will change to invisible in game later on).

Added basic VFX of the corruption in the pound.

Imported a new material from the Unreal 4 examples to use for VFX.